+++ linkcheck/htmlutil/htmlsoup.py
@@ -15,6 +15,35 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 """
 HTML parser implemented using Beautiful Soup and html.parser.
+
+USAGE
+
+Two functions are provided, one to make a BeautifulSoup object from markup and
+another to call a handler's callbacks for each element in a BeautifulSoup
+object it can process.
+
+The used callback of a handler is:
+
+- Start tag: <tag {attr1:value1, attr2:value2, ..}>
+  def start_element (tag, attrs, text, line, column)
+  @param tag: tag name
+  @type tag: string
+  @param attrs: tag attributes
+  @type attrs: dict
+  @param text: element text
+  @type tag: string
+  @param line: tag line number
+  @type tag: integer
+  @param column: tag column number
+  @type tag: integer
+
+EXAMPLE
+
+ # Create a new BeautifulSoup object.
+ soup = htmlutil.htmlsoup.make_soup("<html><body>Blubb</body></html>")
+ # Process the soup with the chosen handler as a parameter.
+ htmlutil.htmlsoup.proces_soup(handler, soup)
+
 """
 
 from warnings import filterwarnings
